backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# AI and ML dependencies
openai==1.3.7
tiktoken==0.5.2

# Vector database
qdrant-client==1.7.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Database (optional - for PostgreSQL)
psycopg2-binary==2.9.9

# Development and testing
pytest==7.4.3
pytest-django==4.7.0 